+++ source/renderer/app/components/staking/redeem-itn-rewards/Step1ConfigurationDialog.js
@@ -143,8 +143,12 @@ type Props = {
   wallets: Array<Wallet>,
 };
 
+type State = {
+  wasRecoveryPhraseValidAtLeastOnce: boolean,
+};
+
 @observer
-export default class Step1ConfigurationDialog extends Component<Props> {
+export default class Step1ConfigurationDialog extends Component<Props, State> {
   static contextTypes = {
     intl: intlShape.isRequired,
   };
@@ -152,6 +156,10 @@ export default class Step1ConfigurationDialog extends Component<Props> {
   static defaultProps = {
     error: null,
     recoveryPhrase: [],
+  };
+
+  state = {
+    wasRecoveryPhraseValidAtLeastOnce: false,
   };
 
   recoveryPhraseAutocomplete: Autocomplete;
@@ -174,6 +182,16 @@ export default class Step1ConfigurationDialog extends Component<Props> {
                 this.context.intl.formatMessage(messages.invalidRecoveryPhrase),
               ],
             }),
+          hooks: {
+            onChange: (field) => {
+              if (
+                this.state.wasRecoveryPhraseValidAtLeastOnce === false &&
+                field.isValid
+              ) {
+                this.setState({ wasRecoveryPhraseValidAtLeastOnce: true });
+              }
+            },
+          },
         },
         walletsDropdown: {
           type: 'select',
@@ -273,16 +291,14 @@ export default class Step1ConfigurationDialog extends Component<Props> {
     const checkboxAcceptance2Field = form.$('checkboxAcceptance2');
     const walletId = get(wallet, 'id', null);
 
-    const validRecoveryPhase = recoveryPhraseField.isValid;
+    const walletsDropdownDisabled = !(
+      recoveryPhraseField.isValid ||
+      this.state.wasRecoveryPhraseValidAtLeastOnce
+    );
 
     const buttonClasses = classnames([
       'primary',
       isCalculatingReedemFees ? styles.isSubmitting : null,
-    ]);
-
-    const walletsDropdownClasses = classnames([
-      styles.walletsDropdown,
-      !validRecoveryPhase ? styles.disabled : null,
     ]);
 
     const actions = {
@@ -371,7 +387,6 @@ export default class Step1ConfigurationDialog extends Component<Props> {
           />
           <div className={styles.walletsDropdownWrapper}>
             <WalletsDropdown
-              className={walletsDropdownClasses}
               {...walletsDropdownField.bind()}
               numberOfStakePools={4}
               wallets={wallets}
@@ -382,6 +397,7 @@ export default class Step1ConfigurationDialog extends Component<Props> {
               value={walletId}
               getStakePoolById={() => {}}
               errorPosition="bottom"
+              disabled={walletsDropdownDisabled}
             />
           </div>
           <Checkbox