+++ source/renderer/app/ipc/getBlockReplayChannel.js
@@ -1,13 +1,13 @@
 // @flow
-import { GET_BLOCK_REPLAY_STATUS_CHANNEL } from '../../../common/ipc/api';
+import { GET_BLOCK_SYNC_STATUS_CHANNEL } from '../../../common/ipc/api';
 import type {
-  GetBlockReplayProgressRendererRequest,
-  GetBlockReplayProgressMainResponse,
+  GetBlockSyncProgressRendererRequest,
+  GetBlockSyncProgressMainResponse,
 } from '../../../common/ipc/api';
 import { RendererIpcChannel } from './lib/RendererIpcChannel';
 
 export const getBlockReplayProgressChannel: // IpcChannel<Incoming, Outgoing>
 RendererIpcChannel<
-  GetBlockReplayProgressMainResponse,
-  GetBlockReplayProgressRendererRequest
-> = new RendererIpcChannel(GET_BLOCK_REPLAY_STATUS_CHANNEL);
+  GetBlockSyncProgressMainResponse,
+  GetBlockSyncProgressRendererRequest
+> = new RendererIpcChannel(GET_BLOCK_SYNC_STATUS_CHANNEL);