+++ source/renderer/app/api/api.js
@@ -94,6 +94,7 @@ import { cardanoFaultInjectionChannel } from '../ipc/cardano.ipc';
 import patchAdaApi from './utils/patchAdaApi';
 import { getLegacyWalletId, utcStringToDate } from './utils';
 import { logger } from '../utils/logging';
+import { hexToString } from '../utils/strings';
 import {
   unscrambleMnemonics,
   scrambleMnemonics,
@@ -2739,6 +2740,7 @@ const _createWalletFromServerData = action(
           uniqueId,
           policyId,
           assetName,
+          assetNameASCII: hexToString(assetName),
           quantity: new BigNumber(quantity.toString()),
         };
       }),
@@ -2749,6 +2751,7 @@ const _createWalletFromServerData = action(
           uniqueId,
           policyId,
           assetName,
+          assetNameASCII: hexToString(assetName),
           quantity: new BigNumber(quantity.toString()),
         };
       }),