+++ source/renderer/app/stores/ProfileStore.js
@@ -43,6 +43,7 @@ import {
   TIME_OPTIONS,
   PROFILE_SETTINGS,
 } from '../config/profileConfig';
+import formatCpuInfo from '../utils/formatCpuInfo';
 
 export default class ProfileStore extends Store {
   @observable systemLocale: Locale = LOCALES.english;
@@ -568,7 +569,7 @@ export default class ProfileStore extends Store {
       const systemInfo = {
         platform: os,
         platformVersion,
-        cpu: Array.isArray(cpu) ? cpu[0].model : '',
+        cpu: formatCpuInfo(cpu),
         ram: formattedBytesToSize(ram),
         availableDiskSpace: diskSpaceAvailable,
       };