+++ storybook/stories/wallets/index.js
@@ -10,6 +10,7 @@ import './transactions/TransactionMetadata.stories';
 import './transactions/Utxo.stories';
 import './tokens/WalletTokens.stories';
 import './tokens/WalletTokensList.stories';
+import '../../../source/renderer/app/components/wallet/tokens/wallet-token-picker/WalletTokenPicker.stories';
 import './settings/WalletSettings.stories';
 import './addWallet/AddWallet.stories';
 import './import/WalletImportFile.stories';